Raspberry Pi Sensors/gps-api/gps_api-master/requirements.txt
pip==20.2.4
bump2version==1.0.1
wheel==0.35.1
watchdog==0.10.3
flake8==3.8.4
tox==3.20.1
coverage==5.3
Sphinx==3.3.0
twine==3.2.0
Click==7.1.2
pytest==6.1.2
pytest-runner==5.2
pyserial==3.4
pynmea2==1.15.0
haversine==2.3.0
